fix - correctly re login on refresh of links view
- custom hook for useAttemptLogin
- can optionally load habit data after the login attempt to refresh
